+++ look/example/lib/pages/widgets/shape.look.dart
@@ -1,0 +1,20 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// LookGenerator
+// **************************************************************************
+
+import 'package:example/pages/widgets/shape.dart';
+import 'package:flutter/material.dart';
+
+main() => runApp(const LookDynamicShapeApp());
+
+class LookDynamicShapeApp extends StatelessWidget {
+  const LookDynamicShapeApp({super.key});
+
+  @override
+  Widget build(BuildContext context) => MaterialApp(
+      title: 'Looking DynamicShape',
+      theme: ThemeData(primarySwatch: Colors.blue),
+      home: dynamicShapeBuilder());
+}